requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
decorator==5.1.1
distlib
Django
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
executing==1.2.0
filelock==3.12.0
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
identify==2.5.24
idna==3.4
inflection==0.5.1
interrogate==1.5.0
ipython==8.13.2
isort==5.12.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pre-commit==3.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
twilio==8.2.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.6
whitenoise==6.4.0
yarl==1.9.2
qrcode